backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
websockets==12.0
python-dotenv==1.0.0 